Properly fix issue with Lure levels (#54)
The previous fix (commit/a9952361d2f255895c6add21c171d42f0b09ebcd)
only applied the special casing for the Lure enchantment in
EnchModule#getEnchInfo and only if the EnchModule.ENCHANTMENT_INFO map
didn't already have an entry for it, however, the map is already
populated in EnchModule#init so the code block that the special casing
is in never executes.

This applies the special casing in EnchModule#init and also applies it
to the config so that the config also show the correct values. (The
config setting is pretty much pointless, but kept for posterity.)

Signed-off-by: srs-bsns <forfrdm@gmail.com>