Update zh_cn.json
Change "core armor" and "lower armor" translations. And there is some personal suggest for who translate this.
The reason is below,
I don't think "core armor" should be translated as "高等盔甲" *(high level armor)*,  and the "lower armor" should be translated as "低等盔甲"*(low level armor)*, 
If only translate "higher armor" and "lower armor", I would translate as "上身盔甲" *(upper body armor)* and "下身盔甲" *(lower body armor)*.
Through my gameplay, the *"core armor"* is *"chestplate"* and *”leggings“*, which should be translated as "核心盔甲"*(core armor)* or "主要盔甲"*(main armor)*, and the *"lower armor"* is *"leggings and boots"*, which should be translted as "下身盔甲"*(lower body armor)*.
The suggestion is below,
And the "tyrannical gem" should be translated as "暴君宝石" instead of "暴政宝石", though the meaning of Trant and Tranny get different preferance. This is somehow “abstract" if a player see "暴政宝石". The ”暴政“ is a low-frequency expression in casul life, we often talk about this in serious. But I won't change this. Literally, this is correct.
As for other translations, I think some of them should be simplified, like "炽热炼狱宝石", just translate as "炽狱宝石", "炽" of "白炽" means *"very high temperature"*, "狱" of "炼狱" means *"inferno"* , 4 characters like older translations of this mod is enough. But this is also personally.